requirements.txt
ultralytics==8.3.179
streamlit>=1.33
opencv-python-headless==4.8.1.78
pillow>=10.0.0
torch==2.2.0
torchvision==0.17.0